Add resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishCommandHandler.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish;
@@ -11,7 +13,8 @@ public class DeleteDishCommandHandler(
     ILogger<DeleteDishCommandHandler> logger,
     IMapper mapper,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository)
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService)
     : IRequestHandler<DeleteDishCommand>
 {
     public async Task Handle(DeleteDishCommand request, CancellationToken cancellationToken)
@@ -23,6 +26,11 @@ public class DeleteDishCommandHandler(
         if (restaurant is null)
         {
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+        }
+        
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbiddenException();
         }
 
         var isDeleted = await dishesRepository.DeleteDishAsync(request.DishId);
